Added hover effect and rounded corners to project images for enhanced visual appeal.
