+++ public/css/main.css
@@ -1,13 +1,3 @@
-/*-----------------------------------------------------------------------------------
-
-    Template Name: Diego - Personal Creative Portfolio & Resume HTML Template
-    Author: Theme_Pure
-    Support: https://help.themepure.net/support/
-    Description: Diego - Personal Creative Portfolio & Resume HTML Template
-    Version: 1.0
-
------------------------------------------------------------------------------------
-
 /************ TABLE OF CONTENTS ***************
 
 	-----------------
@@ -10902,7 +10892,7 @@ xclip
   margin-bottom: 65px;
 }
 .blog-list__tab-btn button {
-  background-color: #272F52;
+  background-color: #272f52;
   border-radius: 25px;
   color: var(--tp-common-white);
   font-family: var(--tp-ff-dmsans);
@@ -11081,7 +11071,7 @@ xclip
   height: 45px;
   width: 45px;
   border-radius: 50%;
-  background-color: #272F52;
+  background-color: #272f52;
   color: var(--tp-common-white);
   transform: 0.3s;
 }
@@ -11393,6 +11383,16 @@ xclip
   .blog-details-4__customize .blog-details__meta {
     margin-bottom: 40px;
   }
+}
+
+.project-display-img {
+  object-fit: cover;
+  object-position: center top;
+  transition: object-position 1.2s ease;
+}
+
+.image-container:hover .project-display-img {
+  object-position: center bottom;
 }
 
 /*----------------------------------------*/
@@ -19868,5 +19868,7 @@ html[tp-theme=tp-theme-light] .tp-menu-2 .tp-main-menu-mobile nav > ul > li a {
   color: var(--tp-common-black);
 }
 html[tp-theme=tp-theme-light] .tp-menu-2 .tp-main-menu-mobile > nav > ul > li.has-dropdown > a .dropdown-toggle-btn svg {
+  color: var(--tp-common-black);
+}svg {
   color: var(--tp-common-black);
 }
