+++ public/scss/theme/_theme.scss
@@ -1,13 +1,4 @@
 @use '../utils' as *;
-/*-----------------------------------------------------------------------------------
-
-    Template Name: Diego - Personal Creative Portfolio & Resume HTML Template
-    Author: Theme_Pure
-    Support: https://help.themepure.net/support/
-    Description: Diego - Personal Creative Portfolio & Resume HTML Template
-    Version: 1.0
-
------------------------------------------------------------------------------------
 
 /************ TABLE OF CONTENTS ***************
